copia y arregla algunas de 2.0, 29 faillures, 25 errors

+++ test/controllers/sip/control_acceso_basicas_controller_test.rb
@@ -1,7 +1,7 @@
 require 'test_helper'
 require 'nokogiri'
 
-module Sivel2Gen
+module Sip
   class ControlAccesoBasicasControllerTest < ActionDispatch::IntegrationTest
 
     include Rails.application.routes.url_helpers
@@ -11,15 +11,10 @@ module Sivel2Gen
       if ENV['CONFIG_HOSTS'] != 'www.example.com'
         raise 'CONFIG_HOSTS debe ser www.example.com'
       end
-      @ope_sin_grupo = Usuario.create!(PRUEBA_USUARIO_OP)
-      @ope_analista = inicia_analista
-    end
-
-    def inicia_analista
-      current_usuario = Usuario.create!(PRUEBA_USUARIO_AN)
-      current_usuario.sip_grupo_ids = [20]
-      current_usuario.save
-      return current_usuario
+      @persona = Sip::Persona.create!(PRUEBA_PERSONA)
+      @ope_sin_grupo = ::Usuario.find(PRUEBA_USUARIO_OP)
+      @ope_analista = ::Usuario.find(PRUEBA_USUARIO_AN)
+      @raiz = Rails.application.config.relative_url_root.delete_suffix('/')
     end
 
     test "sin autenticar no debe listar tablas básicas" do
@@ -29,32 +24,32 @@ module Sivel2Gen
       assert(filas_index == 0)
     end
 
-    basicas_sivel2_gen = Sivel2Gen::Ability::BASICAS_PROPIAS
+    basicas_sip = Sip::Ability::BASICAS_PROPIAS
 
     ## PROBANDO BASICAS GEOGRÁFICAS
+    PAIS_PARAMS = {id: 1, nombre: "ejemplo", nombreiso: "eje", fechacreacion: "2021-12-09"}
     MODELO_PARAMS = {nombre: "ejemplop",observaciones: "obs", fechacreacion: "2021-12-09"}
-    MODELO_PARAMS_IDSTR = { id: "z", nombre: "ejemplop", observaciones: "obs", fechacreacion: "2021-12-09"}
-   
+    MODELO_PARAMS_IDSTR = { id: "a", nombre: "ejemplop", observaciones: "obs", fechacreacion: "2021-12-09"}
+
     def crear_registro(modelo, basica)
       if modelo.columns_hash['id'].type == "string".to_sym
-        case basica
-        when 'trelacion'
+        if basica == 'trelacion'
           registro = modelo.create!(MODELO_PARAMS_IDSTR.merge({inverso: "a"}))
-        when 'tviolencia'
-          registro = modelo.create!(MODELO_PARAMS_IDSTR.merge({nomcorto: "nc"}))
         else
           registro = modelo.create!(MODELO_PARAMS_IDSTR)
         end
       else
         case basica
-        when "categoria"
-          registro = modelo.create!(MODELO_PARAMS.merge({id: 1000, supracategoria_id: 9}))
-        when "pconsolidado"
-          registro = modelo.create!(MODELO_PARAMS.merge({tipoviolencia: "D", clasificacion: "clas"}))
-        when "intervalo"
-          registro = modelo.create!(MODELO_PARAMS.merge({rango: "SIN INFORMACIÓN"}))
-        when "supracategoria"
-          registro = modelo.create!(MODELO_PARAMS.merge({id: 1360, id_tviolencia: "D"}))
+        when "pais"
+          registro = modelo.create!(MODELO_PARAMS.merge({id: 1000, nombreiso_espanol: "iso"}))
+        when "departamento"
+          registro = modelo.create!(MODELO_PARAMS.merge({id_pais: 170, id_deplocal:100000}))
+        when "municipio"
+          registro = modelo.create!(MODELO_PARAMS.merge({id_departamento: 17, id_munlocal:10000}))
+        when "clase"
+          registro = modelo.create!(MODELO_PARAMS.merge({id_municipio: 1360, id_clalocal:10000}))
+        when "vereda"
+          registro = modelo.create!(MODELO_PARAMS.merge({municipio_id: 1360}))
         else
           registro = modelo.create!(MODELO_PARAMS)
         end
@@ -62,9 +57,8 @@ module Sivel2Gen
       return registro
     end
 
-    basicas_sivel2_gen.each do |basica|
-      if basica[1] == "estadocivil" || basica[1] == "maternidad" || basica[1] == "actividadoficio" || basica[1] == "escolaridad"
-
+    basicas_sip.each do |basica|
+      if basica[1] == "oficina"
         next
       end
 
@@ -74,27 +68,40 @@ module Sivel2Gen
 
       #No autenticado
 
-      test "sin autenticar no debe presentar el index de #{basica[1]}" do
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+      if basica[1] == "clase" || basica[1] == "municipio" || basica[1] == "departamento" || basica[1] == "pais"
+        test "sin autenticar debe presentar el index de #{basica[1]}" do
+          get @raiz + "/admin/#{basica[1].pluralize()}"
+          assert_response :ok
         end
-      end
-      test "sin autenticar no debe presentar el show de #{basica[1]}" do
-        reg = crear_registro(modelo, basica[1])
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+        test "sin autenticar debe presentar el show de #{basica[1]}" do
+          skip 
+          reg = modelo.all.take
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          assert_response :ok
         end
-        reg.destroy!
+      else 
+        test "sin autenticar no debe presentar el index de #{basica[1]}" do
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}"
+          end
+        end
+        test "sin autenticar no debe presentar el show de #{basica[1]}" do
+          reg = crear_registro(modelo, basica[1])
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          end
+          reg.destroy!
+        end
       end
 
       test "sin autenticar no debe ver formulario de nuevo de #{basica[1]}" do
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/nueva"
+          get @raiz + "/admin/#{basica[1].pluralize()}/nueva"
         end
       end
 
       test "sin autenticar no puede crear registro de #{basica[1]}" do
-        ruta = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+        ruta = @raiz + "/admin/#{basica[1].pluralize()}"
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
           post ruta, params: {"#{basica[1]}": reg.attributes} 
@@ -105,7 +112,7 @@ module Sivel2Gen
       test "sin autenticar no debe editar #{basica[1]}" do
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}/edita"
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}/edita"
         end
         reg.destroy!
       end
@@ -113,14 +120,14 @@ module Sivel2Gen
       test "sin autenticar no debe actualizar #{basica[1]}" do
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          patch ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+          patch @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
         end
         reg.destroy!
       end
 
       test "sin autenticar no debe dejar destruir un registro de #{basica[1]}" do
         reg = crear_registro(modelo, basica[1])
-        ruta1 = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
+        ruta1 = @raiz + "/admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
         assert_raise CanCan::AccessDenied do
           delete ruta1
         end
@@ -131,31 +138,46 @@ module Sivel2Gen
 
       # Autenticado como operador sin grupo
 
-      test "operador sin grupo no debe presentar el index de #{basica[1]}" do
-        sign_in @ope_sin_grupo
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+      if basica[1] == "clase" || basica[1] == "municipio" || basica[1] == "departamento" || basica[1] == "pais"
+        test "operador sin grupo debe presentar el index de #{basica[1]}" do
+          sign_in @ope_sin_grupo
+          get @raiz + "/admin/#{basica[1].pluralize()}"
+          assert_response :ok
         end
-      end
-      test "operador sin grupo no debe presentar el show de #{basica[1]}" do
-        sign_in @ope_sin_grupo
-        reg = crear_registro(modelo, basica[1])
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+        test "operador sin grupo debe presentar el show de #{basica[1]}" do
+          skip 
+          sign_in @ope_sin_grupo
+          reg = modelo.all.take
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          assert_response :ok
         end
-        reg.destroy!
+      else 
+        test "operador sin grupo no debe presentar el index de #{basica[1]}" do
+          sign_in @ope_sin_grupo
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}"
+          end
+        end
+        test "operador sin grupo no debe presentar el show de #{basica[1]}" do
+          sign_in @ope_sin_grupo
+          reg = crear_registro(modelo, basica[1])
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          end
+          reg.destroy!
+        end
       end
 
       test "operador sin grupo no debe ver formulario de nuevo de #{basica[1]}" do
         sign_in @ope_sin_grupo
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/nueva"
+          get @raiz + "/admin/#{basica[1].pluralize()}/nueva"
         end
       end
 
       test "operador sin grupo no puede crear registro de #{basica[1]}" do
         sign_in @ope_sin_grupo
-        ruta = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+        ruta = @raiz + "/admin/#{basica[1].pluralize()}"
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
           post ruta, params: {"#{basica[1]}": reg.attributes} 
@@ -167,7 +189,7 @@ module Sivel2Gen
         sign_in @ope_sin_grupo
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}/edita"
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}/edita"
         end
         reg.destroy!
       end
@@ -176,7 +198,7 @@ module Sivel2Gen
         sign_in @ope_sin_grupo
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          patch ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+          patch @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
         end
         reg.destroy!
       end
@@ -184,7 +206,7 @@ module Sivel2Gen
       test "oeprador sin grupo no debe dejar destruir un registro de #{basica[1]}" do
         sign_in @ope_sin_grupo
         reg = crear_registro(modelo, basica[1])
-        ruta1 = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
+        ruta1 = @raiz + "/admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
         assert_raise CanCan::AccessDenied do
           delete ruta1
         end
@@ -194,31 +216,46 @@ module Sivel2Gen
 
       # Autenticado como operador con grupo Analista de Casos
 
-      test "operador analista no debe presentar el index de #{basica[1]}" do
-        sign_in @ope_analista
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+      if basica[1] == "clase" || basica[1] == "municipio" || basica[1] == "departamento" || basica[1] == "pais"
+        test "operador analista debe presentar el index de #{basica[1]}" do
+          sign_in @ope_analista
+          get @raiz + "/admin/#{basica[1].pluralize()}"
+          assert_response :ok
         end
-      end
-      test "operador analista no debe presentar el show de #{basica[1]}" do
-        sign_in @ope_analista
-        reg = crear_registro(modelo, basica[1])
-        assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+        test "operador analista debe presentar el show de #{basica[1]}" do
+          skip 
+          sign_in @ope_analista
+          reg = modelo.all.take
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          assert_response :ok
         end
-        reg.destroy!
+      else 
+        test "operador analista no debe presentar el index de #{basica[1]}" do
+          sign_in @ope_analista
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}"
+          end
+        end
+        test "operador analista no debe presentar el show de #{basica[1]}" do
+          sign_in @ope_analista
+          reg = crear_registro(modelo, basica[1])
+          assert_raise CanCan::AccessDenied do
+            get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
+          end
+          reg.destroy!
+        end
       end
 
       test "operador analista no debe ver formulario de nuevo de #{basica[1]}" do
         sign_in @ope_analista
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/nueva"
+          get @raiz + "/admin/#{basica[1].pluralize()}/nueva"
         end
       end
 
       test "operador analista no puede crear registro de #{basica[1]}" do
         sign_in @ope_analista
-        ruta = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}"
+        ruta = @raiz + "/admin/#{basica[1].pluralize()}"
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
           post ruta, params: {"#{basica[1]}": reg.attributes} 
@@ -230,7 +267,7 @@ module Sivel2Gen
         sign_in @ope_analista
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          get ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}/edita"
+          get @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}/edita"
         end
         reg.destroy!
       end
@@ -239,7 +276,7 @@ module Sivel2Gen
         sign_in @ope_analista
         reg = crear_registro(modelo, basica[1])
         assert_raise CanCan::AccessDenied do
-          patch ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}/#{reg.id}"
+          patch @raiz + "/admin/#{basica[1].pluralize()}/#{reg.id}"
         end
         reg.destroy!
       end
@@ -247,7 +284,7 @@ module Sivel2Gen
       test "oeprador analista no debe dejar destruir un registro de #{basica[1]}" do
         sign_in @ope_analista
         reg = crear_registro(modelo, basica[1])
-        ruta1 = ENV['RUTA_RELATIVA'] + "admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
+        ruta1 = @raiz + "/admin/#{basica[1].pluralize()}" + "/" + reg.id.to_s
         assert_raise CanCan::AccessDenied do
           delete ruta1
         end
